trivial: InfoHandler is now standalone class
