+++ src/main/java/org/sysethereum/agents/core/RestServer.java
@@ -1,6 +1,5 @@
 package org.sysethereum.agents.core;
 
-import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpServer;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
@@ -9,7 +8,6 @@ import org.springframework.stereotype.Service;
 import org.sysethereum.agents.service.rest.*;
 
 import javax.annotation.PostConstruct;
-import java.io.IOException;
 import java.net.InetSocketAddress;
 
 /**
@@ -28,23 +26,26 @@ public class RestServer {
     private final GetSuperblockBySyscoinHandler getSuperblockBySyscoinHandler;
     private final GetSuperblockHandler getSuperblockHandler;
     private final GetSyscoinRPCHandler getSyscoinRPCHandler;
+    private final InfoHandler infoHandler;
 
     public RestServer(
             GetSPVHandler getSPVHandler,
             GetSuperblockBySyscoinHandler getSuperblockBySyscoinHandler,
             GetSuperblockHandler getSuperblockHandler,
-            GetSyscoinRPCHandler getSyscoinRPCHandler
+            GetSyscoinRPCHandler getSyscoinRPCHandler,
+            InfoHandler infoHandler
     ) {
         this.getSPVHandler = getSPVHandler;
         this.getSuperblockBySyscoinHandler = getSuperblockBySyscoinHandler;
         this.getSuperblockHandler = getSuperblockHandler;
         this.getSyscoinRPCHandler = getSyscoinRPCHandler;
+        this.infoHandler = infoHandler;
     }
 
     @PostConstruct
     public void setup() throws Exception {
         HttpServer server = HttpServer.create(new InetSocketAddress(8000), 0);
-        server.createContext("/", new InfoHandler());
+        server.createContext("/", infoHandler);
         server.createContext("/spvproof", getSPVHandler);
         server.createContext("/superblockbysyscoinblock", getSuperblockBySyscoinHandler);
         server.createContext("/superblock", getSuperblockHandler);
@@ -52,24 +53,4 @@ public class RestServer {
         server.setExecutor(null); // creates a default executor
         server.start();
     }
-
-    // http://localhost:8000/info
-    private static class InfoHandler extends CommonHttpHandler {
-
-        @Override
-        public void handle(HttpExchange httpExchange) throws IOException {
-            String response = "Valid Superblock calls: " + System.lineSeparator() +
-                    "\t/spvproof?hash=<blockhash>" + System.lineSeparator() +
-                    "\t/spvproof?height=<blockheight>" + System.lineSeparator() +
-                    "\t/superblockbysyscoinblock?hash=<blockhash>" + System.lineSeparator() +
-                    "\t/superblockbysyscoinblock?height=<blockheight>" + System.lineSeparator() +
-                    "\t/superblock?hash=<superblockid>" + System.lineSeparator() +
-                    "\t/superblock?height=<superblockheight>" + System.lineSeparator() + System.lineSeparator() +
-                    "Valid Syscoin RPC calls: " + System.lineSeparator() +
-                    "\t/syscoinrpc?method=<methodname>&param1name=<param1value>&paramNname=<paramNvalue>...";
-            writeResponse(httpExchange, response);
-        }
-    }
-
 }
-
